+++ lib/config/routes/routes.dart
@@ -9,10 +9,10 @@ final _routes = <RouteBase>[
     },
     routes: [
       GoRoute(
-        path: SearchProductsPage.routeName, // 'home/search'
+        path: SearchPage.routeName, // 'home/search'
         builder: (context, state) {
           final String keywords = state.extra as String;
-          return SearchProductsPage(keywords: keywords);
+          return SearchPage(keywords: keywords);
         },
         // routes: [
         //   GoRoute(
@@ -28,9 +28,9 @@ final _routes = <RouteBase>[
   ),
   // User Route
   GoRoute(
-    path: '/${UserHome.routeName}', // '/user'
+    path: '/${UserHomePage.routeName}', // '/user'
     builder: (BuildContext context, GoRouterState state) {
-      return const UserHome(); // contain login page
+      return const UserHomePage(); // contain login page
     },
     routes: [
       GoRoute(