+++ lib/features/home/data/models/category_model.dart
@@ -62,10 +62,12 @@ class CategoryModel extends CategoryEntity {
     );
   }
 
-  factory CategoryModel.fromJson(String source) => CategoryModel.fromMap(json.decode(source) as Map<String, dynamic>);
+  factory CategoryModel.fromJson(String source) =>
+      CategoryModel.fromMap(json.decode(source) as Map<String, dynamic>);
 
   static List<CategoryModel> fromJsonToList(String source) {
-    final Map<String, dynamic> sourceMap = json.decode(source) as Map<String, dynamic>;
+    final Map<String, dynamic> sourceMap =
+        json.decode(source) as Map<String, dynamic>;
     final List<dynamic> mapList = sourceMap['categoryDTOs'] as List<dynamic>;
 
     return mapList.map((e) => CategoryModel.fromMap(e)).toList();
@@ -73,7 +75,8 @@ class CategoryModel extends CategoryEntity {
 
   static List<CategoryModel> fromMapToList(Map<String, dynamic> mapList) {
     return (mapList['categoryDTOs'] as List<dynamic>)
-        .map<CategoryModel>((e) => CategoryModel.fromMap(e as Map<String, dynamic>))
+        .map<CategoryModel>(
+            (e) => CategoryModel.fromMap(e as Map<String, dynamic>))
         .toList();
   }
 }