+++ lib/features/home/domain/entities/product_entity.dart
@@ -67,14 +67,17 @@ class ProductEntity extends Equatable {
       status: map['status'] as String,
       categoryId: map['categoryId'].toInt() as int,
       brandId: map['brandId']?.toInt() as int?,
-      productVariant: ProductVariantEntity.fromList(map['productVariantDTOs'] as List<dynamic>),
+      productVariant: ProductVariantEntity.fromList(
+          map['productVariantDTOs'] as List<dynamic>),
     );
   }
 
-  factory ProductEntity.fromJson(String source) => ProductEntity.fromMap(json.decode(source) as Map<String, dynamic>);
+  factory ProductEntity.fromJson(String source) =>
+      ProductEntity.fromMap(json.decode(source) as Map<String, dynamic>);
 
   static List<ProductEntity> fromList(List<dynamic> list) {
-    return List<ProductEntity>.from(list.map((e) => ProductEntity.fromMap(e as Map<String, dynamic>)));
+    return List<ProductEntity>.from(
+        list.map((e) => ProductEntity.fromMap(e as Map<String, dynamic>)));
   }
 
   @override
@@ -115,11 +118,15 @@ class ProductEntity extends Equatable {
 
   // get cheapest price of the product in the list of product variants
   int get cheapestPrice {
-    return productVariant.map((e) => e.price).reduce((value, element) => value < element ? value : element);
+    return productVariant
+        .map((e) => e.price)
+        .reduce((value, element) => value < element ? value : element);
   }
 
   // get the most expensive price of the product in the list of product variants
   int get mostExpensivePrice {
-    return productVariant.map((e) => e.price).reduce((value, element) => value > element ? value : element);
+    return productVariant
+        .map((e) => e.price)
+        .reduce((value, element) => value > element ? value : element);
   }
 }