+++ lib/features/home/presentation/components/product_components/product_list_builder.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../../../../core/constants/typedef.dart';
-import '../../domain/dto/product_dto.dart';
+import '../../../../../core/constants/typedef.dart';
+import '../../../domain/response/product_resp.dart';
 import 'page_number.dart';
 import 'product_item.dart';
 
@@ -20,8 +20,8 @@ class ProductListBuilder extends StatelessWidget {
   })  : assert(crossAxisCount > 0),
         assert(showPageNumber == false || (currentPage != null && onPageChanged != null));
 
-  final Future<RespData<ProductDTO>> future;
-  final String? keywords;
+  final Future<RespData<ProductResp>> future;
+  final String? keywords; // for search page
   final int crossAxisCount;
 
   // for showing page number component at the bottom
@@ -31,7 +31,7 @@ class ProductListBuilder extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return FutureBuilder<RespData<ProductDTO>>(
+    return FutureBuilder<RespData<ProductResp>>(
       future: future,
       builder: (context, snapshot) {
         if (snapshot.connectionState != ConnectionState.done) {
@@ -46,10 +46,10 @@ class ProductListBuilder extends StatelessWidget {
             ),
             (dataResp) => Builder(builder: (context) {
               if (dataResp.data.products.isEmpty) {
-                return Center(
+                return const Center(
                   child: Text(
-                    keywords == null ? 'Không tìm thấy sản phẩm nào' : 'Không tìm thấy sản phẩm nào cho từ khóa:  "$keywords"',
-                    style: const TextStyle(color: Colors.red),
+                    'Không tìm thấy sản phẩm phù hợp',
+                    style: TextStyle(color: Colors.red),
                   ),
                 );
               }