+++ lib/features/home/presentation/components/search_components/bottom_sheet_filter.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 import '../../../../../core/helpers/helpers.dart';
-import '../../pages/search_products_page.dart';
-import 'btn_search_types.dart';
+import 'btn_filter.dart';
+import 'btn_dropdown_sort_types.dart';
 
 class BottomSheetFilter extends StatefulWidget {
   /// default range is 0 - 10tr
@@ -15,6 +15,7 @@ class BottomSheetFilter extends StatefulWidget {
     this.minRange = 0,
     this.maxRange = 10000000, // 10tr
     this.divisions = 100,
+    this.filterPriceRange = true,
   });
   // required
   final BuildContext context;
@@ -26,6 +27,7 @@ class BottomSheetFilter extends StatefulWidget {
   final double minRange;
   final double maxRange;
   final int divisions;
+  final bool filterPriceRange;
 
   @override
   State<BottomSheetFilter> createState() => _BottomSheetFilterState();
@@ -36,6 +38,9 @@ class _BottomSheetFilterState extends State<BottomSheetFilter> {
   late double _minRange;
   late double _maxRange;
   late String _sortType;
+
+  // optional
+  late bool _filterPriceRange;
 
   @override
   void initState() {
@@ -57,26 +62,54 @@ class _BottomSheetFilterState extends State<BottomSheetFilter> {
       widget.maxPrice.toDouble(),
     );
 
+    _filterPriceRange = widget.filterPriceRange;
     _sortType = widget.sortType;
   }
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+      padding: const EdgeInsets.all(12),
+      height: MediaQuery.of(context).size.height * 0.8,
       child: Column(
         children: [
-          _filterByPrice(),
-          BtnSortTypes(
-            initValue: _sortType,
-            onSortChanged: (sortType) {
-              // don't need setState because BtnSortTypes is a stateful widget itself
-              _sortType = sortType;
-            },
+          const Text(
+            'Lọc',
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.bold,
+            ),
+            textAlign: TextAlign.center,
           ),
+          const Divider(thickness: 0.5, color: Colors.grey),
+          _filterByPrice(),
+          _buildSortTypes(),
           _btnApplyCancel(context),
         ],
       ),
+    );
+  }
+
+  Widget _buildSortTypes() {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        const Text(
+          'Sắp xếp theo',
+          style: TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        const SizedBox(height: 10),
+        BtnDropdownSortTypes(
+          initValue: _sortType,
+          onSortChanged: (sortType) {
+            // don't need setState because BtnSortTypes is a stateful widget itself
+            _sortType = sortType;
+          },
+        ),
+      ],
     );
   }
 
@@ -96,6 +129,7 @@ class _BottomSheetFilterState extends State<BottomSheetFilter> {
                     minPrice: _currentRangeValues.start.round(),
                     maxPrice: _currentRangeValues.end.round(),
                     sortType: _sortType,
+                    filterPriceRange: _filterPriceRange,
                   ),
                 );
               },
@@ -119,6 +153,7 @@ class _BottomSheetFilterState extends State<BottomSheetFilter> {
                       minPrice: _currentRangeValues.start.round(),
                       maxPrice: _currentRangeValues.end.round(),
                       sortType: _sortType,
+                      filterPriceRange: _filterPriceRange,
                     ));
               },
               child: Container(
@@ -137,17 +172,110 @@ class _BottomSheetFilterState extends State<BottomSheetFilter> {
     );
   }
 
+  Widget _buildSuggestionFilterPrice() {
+    return Wrap(
+      spacing: 10,
+      children: [
+        TextButton(
+          style: TextButton.styleFrom(backgroundColor: Colors.grey[300]),
+          onPressed: () {
+            setState(() {
+              _currentRangeValues = const RangeValues(0, 100000);
+              if (!_filterPriceRange) {
+                _filterPriceRange = true;
+              }
+            });
+          },
+          child: const Text('Dưới 100.000'),
+        ),
+        TextButton(
+          style: TextButton.styleFrom(backgroundColor: Colors.grey[300]),
+          onPressed: () {
+            setState(() {
+              _currentRangeValues = const RangeValues(0, 500000);
+              if (!_filterPriceRange) {
+                _filterPriceRange = true;
+              }
+            });
+          },
+          child: const Text('Dưới 500.000'),
+        ),
+        TextButton(
+          style: TextButton.styleFrom(backgroundColor: Colors.grey[300]),
+          onPressed: () {
+            setState(() {
+              _currentRangeValues = const RangeValues(500000, 1000000);
+              if (!_filterPriceRange) {
+                _filterPriceRange = true;
+              }
+            });
+          },
+          child: const Text('500.000 - 1tr'),
+        ),
+        TextButton(
+          style: TextButton.styleFrom(backgroundColor: Colors.grey[300]),
+          onPressed: () {
+            setState(() {
+              _currentRangeValues = const RangeValues(1000000, 2000000);
+              if (!_filterPriceRange) {
+                _filterPriceRange = true;
+              }
+            });
+          },
+          child: const Text('1tr - 2tr'),
+        ),
+        TextButton(
+          style: TextButton.styleFrom(backgroundColor: Colors.grey[300]),
+          onPressed: () {
+            setState(() {
+              _currentRangeValues = const RangeValues(2000000, 5000000);
+              if (!_filterPriceRange) {
+                _filterPriceRange = true;
+              }
+            });
+          },
+          child: const Text('2tr - 5tr'),
+        ),
+        TextButton(
+          style: TextButton.styleFrom(backgroundColor: Colors.grey[300]),
+          onPressed: () {
+            setState(() {
+              _currentRangeValues = const RangeValues(5000000, 10000000);
+              if (!_filterPriceRange) {
+                _filterPriceRange = true;
+              }
+            });
+          },
+          child: const Text('Trên 5tr'),
+        ),
+      ],
+    );
+  }
+
   Column _filterByPrice() {
     return Column(
-      mainAxisSize: MainAxisSize.min,
       children: [
-        const Text(
-          'Lọc theo giá',
-          style: TextStyle(
-            fontSize: 18,
-            fontWeight: FontWeight.bold,
-          ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.start,
+          children: [
+            Checkbox(
+              value: _filterPriceRange,
+              onChanged: (bool? value) {
+                setState(() {
+                  _filterPriceRange = value!;
+                });
+              },
+            ),
+            const Text(
+              'Hiện thị sản phẩm theo giá',
+              style: TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ],
         ),
+        _buildSuggestionFilterPrice(),
         const SizedBox(height: 10),
         Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,